requirements.txt
PyYAML
aiohttp
bs4
fastapi
pandas
fastapi
starlette
openpyxl
uvicorn
jinja2
python-multipart
requests
lxml
html2text
pyinstaller
selenium
webdriver_manager
openpyxl
readability-lxml
lxml_html_clean
